global-requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
autobahn==25.10.2
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.7.14
cffi==2.0.0
charset-normalizer==3.4.2
click==8.3.0
cmdstanpy==1.2.5
colorama==0.4.6
comm==0.2.2
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
dotenv==0.9.9
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
flatbuffers==25.2.10
fonttools==4.59.1
fpdf==1.7.2
fqdn==1.5.1
frozenlist==1.8.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
holidays==0.79
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
importlib_resources==6.5.2
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_proxy==4.4.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_vpython==3.1.8
jupyterlab_widgets==3.0.15
keras==3.11.2
kiwisolver==1.4.9
lark==1.2.2
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.3
multidict==6.7.0
namex==0.1.0
narwhals==2.1.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nbimporter==0.3.4
nest-asyncio==1.6.0
notebook==7.4.4
notebook_shim==0.2.4
numpy==2.3.2
opencage==3.2.0
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.17.0
overrides==7.7.0
packaging==25.0
Panda3D==1.10.15
panda3d-gltf==1.3.0
panda3d-simplepbr==0.13.1
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pillow==11.3.0
platformdirs==4.3.8
plotly==6.3.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.4.1
prophet==1.1.7
protobuf==6.32.0
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==3.3.0
pytz==2025.2
pywin32==311
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.26.0
scikit-learn==1.7.1
scipy==1.16.1
screeninfo==0.8.1
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==80.9.0
simpervisor==1.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
sswg==4.3.0
stack-data==0.6.3
stanio==0.5.1
statsmodels==0.14.5
streamlit==1.49.1
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
txaio==25.9.2
types-python-dateutil==2.9.0.20250708
typing_extensions==4.14.1
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
ursina==8.2.0
vpython==7.6.5
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
wrapt==1.17.3
xlsxwriter==3.2.5
yarl==1.22.0
